+++ frontend/components/commons/forms/questions-form/fields/rating/Rating.component.vue
@@ -1,12 +1,19 @@
 <template>
-  <div class="wrapper">
-    <QuestionHeaderComponent
-      :question="question"
-      :showSuggestion="showSuggestion"
-    />
+  <RatingShortcuts>
+    <div class="wrapper">
+      <QuestionHeaderComponent
+        :question="question"
+        :showSuggestion="showSuggestion"
+      />
 
-    <RatingMonoSelectionComponent v-model="question.answer.values" />
-  </div>
+      <RatingMonoSelectionComponent
+        ref="ratingMonoSelectionRef"
+        v-model="question.answer.values"
+        :isFocused="isFocused"
+        @on-focus="$emit('on-focus')"
+      />
+    </div>
+  </RatingShortcuts>
 </template>
 
 <script>
@@ -20,6 +27,19 @@ export default {
     showSuggestion: {
       type: Boolean,
       default: () => false,
+    },
+    isFocused: {
+      type: Boolean,
+      default: () => false,
+    },
+  },
+  watch: {
+    "question.answer.values": {
+      deep: true,
+      handler(newOptions) {
+        if (newOptions.some((option) => option.isSelected))
+          this.$emit("on-user-answer");
+      },
     },
   },
 };