+++ frontend/components/feedback-task/ContentEditableFeedbackTask.vue
@@ -21,11 +21,16 @@
         @keydown.shift.enter.stop=""
         @keydown.shift.backspace.stop=""
         @keydown.shift.space.stop=""
-        @keydown.esc.exact="looseFocus"
-        @keydown.arrow-right.stop=""
-        @keydown.arrow-left.stop=""
-        @keydown.delete.exact.stop=""
-        @keydown.enter.exact.stop=""
+        @keydown.shift.arrow-up.stop=""
+        @keydown.shift.arrow-down.stop=""
+        @keydown.ctrl.arrow-right.stop=""
+        @keydown.meta.arrow-right.stop=""
+        @keydown.ctrl.arrow-left.stop=""
+        @keydown.meta.arrow-left.stop=""
+        @keydown.delete.stop=""
+        @keydown.enter.stop=""
+        @keydown.esc.exact="exitEditionMode"
+        @paste="pastePlainText"
       />
     </div>
   </span>
@@ -43,13 +48,23 @@ export default {
       type: String,
       default: "",
     },
+    originalValue: {
+      type: String,
+      default: "",
+    },
+    isFocused: {
+      type: Boolean,
+      default: () => false,
+    },
+    isEditionModeActive: {
+      type: Boolean,
+      required: true,
+    },
   },
   data: () => {
     return {
-      originalValue: null,
       sanitizedCurrentValue: null,
       currentValue: null,
-      focus: false,
     };
   },
   computed: {
@@ -58,6 +73,21 @@ export default {
     },
   },
   watch: {
+    isFocused: {
+      immediate: true,
+      handler(newValue) {
+        if (newValue) {
+          this.$nextTick(() => {
+            this.$refs.text.focus();
+          });
+        }
+      },
+    },
+    isEditionModeActive() {
+      if (this.isFocused && this.isEditionModeActive) {
+        this.$refs.text.focus();
+      }
+    },
     value() {
       if (this.value !== this.currentValue) {
         this.reset();
@@ -65,34 +95,30 @@ export default {
     },
   },
   mounted() {
-    window.addEventListener("paste", this.pastePlainText);
-
     this.reset();
-
-    this.textAreaWrapper = document.getElementById("contentId");
-  },
-  destroyed() {
-    window.removeEventListener("paste", this.pastePlainText);
   },
   methods: {
     reset() {
-      this.currentValue = this.originalValue = this.value;
+      this.currentValue = this.value;
       this.sanitizedCurrentValue = " ";
-      this.sanitizedCurrentValue = this.currentValue;
+      this.$nextTick(() => {
+        this.sanitizedCurrentValue = this.currentValue;
+      });
     },
-    looseFocus() {
-      this.textAreaWrapper.blur();
+    exitEditionMode() {
+      this.$refs.text.blur();
+
+      this.$emit("on-exit-edition-mode");
     },
     onInputText(event) {
       this.currentValue = event.target.innerText;
       this.$emit("change-text", this.currentValue);
     },
-    setFocus(status) {
-      this.focus = status;
-      this.$emit("on-change-focus", this.focus);
+    setFocus(isFocus) {
+      this.$emit("on-change-focus", isFocus);
     },
     pastePlainText(event) {
-      if (this.focus && event.target.isContentEditable) {
+      if (event.target.isContentEditable) {
         event.preventDefault();
         const text = event.clipboardData?.getData("text/plain") ?? "";
         document.execCommand("insertText", false, text);
@@ -106,6 +132,7 @@ export default {
 </script>
 <style lang="scss" scoped>
 [contenteditable="true"] {
+  padding: 0.6em;
   outline: none;
   &:focus + span {
     display: block;